Refactor DTOs and enhance security mechanisms
Refactored DTOs to simplify commands and queries, replacing complex objects with individual properties. Renamed `HashToken` to `GetHash` for clarity and updated its usage across the codebase. Introduced `IHttpContextAccessor` to dynamically retrieve `IpAddress` and `UserAgent`.

Added an optional `Password` property to the `UserProfile` entity and created a migration to reflect this change in the database. Updated endpoints in `AccountController` to align with the new DTO structure. Improved maintainability, security, and code clarity.